+++ COVERAGE.md
@@ -1,10 +1,18 @@
-| Name                              |    Stmts |     Miss |   Branch |   BrPart |   Cover |
-|---------------------------------- | -------: | -------: | -------: | -------: | ------: |
-| src/hatch\_cython/\_\_init\_\_.py |        2 |        0 |        0 |        0 |    100% |
-| src/hatch\_cython/config.py       |      283 |       19 |      124 |       12 |     92% |
-| src/hatch\_cython/devel.py        |        5 |        0 |        0 |        0 |    100% |
-| src/hatch\_cython/hooks.py        |        5 |        1 |        2 |        0 |     86% |
-| src/hatch\_cython/plugin.py       |      193 |        9 |      140 |        7 |     95% |
-| src/hatch\_cython/types.py        |       19 |        5 |        2 |        1 |     71% |
-| src/hatch\_cython/utils.py        |       10 |        0 |        2 |        0 |    100% |
-|                         **TOTAL** |  **517** |   **34** |  **270** |   **20** | **93%** |
+| Name                                     |    Stmts |     Miss |   Branch |   BrPart |   Cover |
+|----------------------------------------- | -------: | -------: | -------: | -------: | ------: |
+| src/hatch\_cython/\_\_init\_\_.py        |        2 |        0 |        0 |        0 |    100% |
+| src/hatch\_cython/config/\_\_init\_\_.py |        2 |        0 |        0 |        0 |    100% |
+| src/hatch\_cython/config/autoimport.py   |        9 |        0 |        4 |        0 |    100% |
+| src/hatch\_cython/config/config.py       |      145 |       16 |       70 |       10 |     87% |
+| src/hatch\_cython/config/defaults.py     |        6 |        0 |        0 |        0 |    100% |
+| src/hatch\_cython/config/files.py        |        5 |        0 |        2 |        0 |    100% |
+| src/hatch\_cython/config/flags.py        |       53 |        1 |       16 |        0 |     99% |
+| src/hatch\_cython/config/platform.py     |       71 |        2 |       28 |        2 |     96% |
+| src/hatch\_cython/constants.py           |       11 |        0 |        0 |        0 |    100% |
+| src/hatch\_cython/devel.py               |        5 |        0 |        0 |        0 |    100% |
+| src/hatch\_cython/hooks.py               |        5 |        1 |        2 |        0 |     86% |
+| src/hatch\_cython/plugin.py              |      186 |        9 |      140 |        8 |     95% |
+| src/hatch\_cython/temp.py                |       12 |        0 |        2 |        0 |    100% |
+| src/hatch\_cython/types.py               |       19 |        4 |        2 |        1 |     76% |
+| src/hatch\_cython/utils.py               |       22 |        0 |        8 |        0 |    100% |
+|                                **TOTAL** |  **553** |   **33** |  **274** |   **21** | **93%** |
