+++ src/hatch_cython/config/autoimport.py
@@ -1,0 +1,26 @@
+from dataclasses import dataclass, field
+
+
+@dataclass
+class Autoimport:
+    pkg: str
+
+    include: str
+    libraries: str = field(default=None)
+    library_dirs: str = field(default=None)
+    required_call: str = field(default=None)
+
+
+__packages__ = {
+    a.pkg: a
+    for a in (
+        Autoimport("numpy", "get_include"),
+        Autoimport(
+            "pyarrow",
+            include="get_include",
+            libraries="get_libraries",
+            library_dirs="get_library_dirs",
+            required_call="create_library_symlinks",
+        ),
+    )
+}
