+++ src/hatch_cython/config/files.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass, field
+
+from hatch_cython.types import ListStr
+
+
+@dataclass
+class FileArgs:
+    exclude: ListStr = field(default_factory=list)
