+++ src/hatch_cython/constants.py
@@ -1,0 +1,15 @@
+from hatch_cython.types import CorePlatforms, list_t
+
+NORM_GLOB = r"([^\s]*)"
+UAST = "${U_AST}"
+EXIST_TRIM = 2
+ANON = "anon"
+INCLUDE = "include_"
+OPTIMIZE = "-O2"
+DIRECTIVES = {
+    "binding": True,
+    "language_level": 3,
+}
+LTPY311 = "python_version < '3.11'"
+MUST_UNIQUE = ["-O", "-arch", "-march"]
+POSIX_CORE: list_t[CorePlatforms] = ["darwin", "linux"]
