+++ src/hatch_cython/utils.py
@@ -1,7 +1,10 @@
-from hatch_cython.types import P, T, callable_t
+import platform
+
+from hatch_cython.constants import NORM_GLOB, UAST
+from hatch_cython.types import CallableT, P, T
 
 
-def memo(func: callable_t[P, T]) -> callable_t[P, T]:
+def memo(func: CallableT[P, T]) -> CallableT[P, T]:
     value = None
     ran = False
 
@@ -13,3 +16,21 @@ def memo(func: callable_t[P, T]) -> callable_t[P, T]:
         return value
 
     return wrapped
+
+
+@memo
+def plat():
+    return platform.system().lower()
+
+
+@memo
+def aarch():
+    return platform.machine().lower()
+
+
+def options_kws(kwds: dict):
+    return ",\n\t".join((f"{k}={v!r}" for k, v in kwds.items()))
+
+
+def parse_user_glob(uglob: str):
+    return uglob.replace("\\*", UAST).replace("*", NORM_GLOB).replace(UAST, "*")
