+++ tests/test_plugin.py
@@ -84,20 +84,21 @@ def test_build_hook(new_proj):
             ]
         )
 
-        assert sorted([sorted(ls) for ls in hook.grouped_included_files]) == sorted(
-            [
-                sorted(ls)
-                for ls in [
-                    ["./src/example_lib/normal.py"],
-                    ["./src/example_lib/__init__.py"],
-                    ["./src/example_lib/__about__.py"],
-                    ["./src/example_lib/mod_a/__init__.py"],
-                    ["./src/example_lib/mod_a/some_defn.py", "./src/example_lib/mod_a/some_defn.py"],
-                    ["./src/example_lib/test.pyx"],
-                    ["./src/example_lib/mod_a/adds.pyx"],
-                ]
-            ]
-        )
+        assert sorted(
+            [{**ls, "files": sorted(ls.get("files"))} for ls in hook.grouped_included_files],
+            key=lambda x: x.get("name"),
+        ) == [
+            {"name": "example_lib.__about__", "files": ["./src/example_lib/__about__.py"]},
+            {"name": "example_lib.__init__", "files": ["./src/example_lib/__init__.py"]},
+            {"name": "example_lib.mod_a.__init__", "files": ["./src/example_lib/mod_a/__init__.py"]},
+            {"name": "example_lib.mod_a.adds", "files": ["./src/example_lib/mod_a/adds.pyx"]},
+            {
+                "name": "example_lib.mod_a.some_defn",
+                "files": ["./src/example_lib/mod_a/some_defn.py", "./src/example_lib/mod_a/some_defn.py"],
+            },
+            {"name": "example_lib.normal", "files": ["./src/example_lib/normal.py"]},
+            {"name": "example_lib.test", "files": ["./src/example_lib/test.pyx"]},
+        ]
 
         rf = sorted(
             [
